backend/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
httpx==0.25.2
